storage: rename segment_appender_utils.*
The utilities are used in the appender, but are used broadly for generic
serialization from disk. This renames the corresponding files to
record_batch_utils.*.
