+++ src/v/cloud_storage/tests/remote_segment_test.cc
@@ -26,8 +26,8 @@
 #include "model/timeout_clock.h"
 #include "storage/log.h"
 #include "storage/log_manager.h"
+#include "storage/record_batch_utils.h"
 #include "storage/segment.h"
-#include "storage/segment_appender_utils.h"
 #include "storage/tests/utils/disk_log_builder.h"
 #include "storage/types.h"
 #include "test_utils/async.h"
