+++ src/v/storage/compaction_reducers.cc
@@ -18,7 +18,7 @@
 #include "storage/index_state.h"
 #include "storage/logger.h"
 #include "storage/parser_utils.h"
-#include "storage/segment_appender_utils.h"
+#include "storage/record_batch_utils.h"
 #include "storage/segment_utils.h"
 
 #include <seastar/core/future.hh>
