+++ src/v/storage/record_batch_utils.h
