+++ src/v/storage/tests/log_replayer_test.cc
@@ -16,8 +16,8 @@
 #include "storage/file_sanitizer.h"
 #include "storage/log_replayer.h"
 #include "storage/logger.h"
+#include "storage/record_batch_utils.h"
 #include "storage/segment.h"
-#include "storage/segment_appender_utils.h"
 #include "storage/segment_index.h"
 #include "storage/segment_reader.h"
 #include "storage/storage_resources.h"
